refactor: PaymentService 삭제
